requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
firebase-admin==6.5.0
python-socketio[asyncio_asgi]==5.11.2
email-validator==2.1.1
